Clean up documentation. Seems to work better to have str first in the vars enum and cast everything to char *.

+++ ansi.h
@@ -52,4 +52,4 @@ struct key_array Tkeys[] = {
 	{ "\033[7^",	"C-home" },
 	{ "\033[8^",	"C-end" },
 };
-#define N_KEYS ((sizeof(Tkeys) / sizeof(struct key_array)))
+#define N_KEYS ((int)(sizeof(Tkeys) / sizeof(struct key_array)))
